fix: resolve CI fmt failure and all audit findings (Rust + TS)
Agent-Logs-Url: https://github.com/el-j/maze-game-esp32/sessions/54c30afa-569b-4067-8295-bdc53e14e7d5

+++ packages/firmware-rust/src/config.rs
@@ -28,10 +28,14 @@ pub const MOTOR_MS_VICTORY: u32 = 600;
 
 // ── Game Rules ───────────────────────────────────────────────
 pub const STARTING_LIVES: i32 = 3;
-pub const CRASH_DISPLAY_MS: u32 = 1_500;
-/// Duration of the non-blocking pause shown between levels.
-pub const LEVELUP_PAUSE_MS: u32 = 1_200;
+/// How long the crash screen is displayed before respawning (ms).
+/// `u64` so it can be added directly to `now_ms` without a cast.
+pub const CRASH_DISPLAY_MS: u64 = 1_500;
+/// Duration of the non-blocking pause shown between levels (ms).
+/// `u64` so it can be added directly to `now_ms` without a cast.
+pub const LEVELUP_PAUSE_MS: u64 = 1_200;
 /// Minimum milliseconds between accepted button presses.
 /// Prevents a single held button from cascading through multiple state
 /// transitions (e.g. GAMEOVER → TITLE → PLAYING in one press).
+/// `u64` so it can be added directly to `now_ms` without a cast.
 pub const DEBOUNCE_MS: u64 = 300;
